+++ src/index.js
@@ -1,4 +1,5 @@
-import App from './App';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './dev/'; 
 
-export default App;
-
+ReactDOM.render(<App />, document.getElementById('root'));